+++ spec/policy/caching/policy_spec.lua
@@ -1,9 +1,21 @@
 local resty_lrucache = require('resty.lrucache')
 
 describe('policy', function()
-  describe('.new', function()
-    local cache = resty_lrucache.new(1)
+  local cache
 
+  before_each(function()
+    cache = resty_lrucache.new(1)
+
+    -- The code uses ngx.shared.dict and it defines .add(), resty_lrucache
+    -- does not, so we need to implement it for these tests.
+    cache.add = function(self, key, value)
+      if not self:get(key) then
+        return self:set(key, value)
+      end
+    end
+  end)
+
+  describe('.new', function()
     it('disables caching when caching type is not specified', function()
       local caching_policy = require('apicast.policy.caching').new({})
       local ctx = {}
@@ -27,7 +39,6 @@ describe('policy', function()
   describe('.access', function()
     describe('when configured as strict', function()
       local caching_policy
-      local cache
       local ctx  -- the caching policy will add the handler here
 
       before_each(function()
@@ -35,7 +46,6 @@ describe('policy', function()
         caching_policy = require('apicast.policy.caching').new(config)
         ctx = { }
         caching_policy:rewrite(ctx)
-        cache = resty_lrucache.new(1)
       end)
 
       it('caches authorized requests', function()
@@ -60,7 +70,6 @@ describe('policy', function()
 
     describe('when configured as resilient', function()
       local caching_policy
-      local cache
       local ctx  -- the caching policy will add the handler here
 
       before_each(function()
@@ -68,7 +77,6 @@ describe('policy', function()
         caching_policy = require('apicast.policy.caching').new(config)
         ctx = { }
         caching_policy:rewrite(ctx)
-        cache = resty_lrucache.new(1)
       end)
 
       it('caches authorized requests', function()
@@ -89,9 +97,49 @@ describe('policy', function()
       end)
     end)
 
+    describe('when configured as allow', function()
+      local caching_policy
+      local ctx  -- the caching policy will add the handler here
+
+      before_each(function()
+        local config = { caching_type = 'allow' }
+        caching_policy = require('apicast.policy.caching').new(config)
+        ctx = { }
+        caching_policy:rewrite(ctx)
+      end)
+
+      it('caches authorized requests', function()
+        ctx.cache_handler(cache, 'a_key', { status = 200 }, nil)
+        assert.equals(200, cache:get('a_key'))
+      end)
+
+      it('caches denied requests', function()
+        ctx.cache_handler(cache, 'a_key', { status = 403 }, nil)
+        assert.equals(403, cache:get('a_key'))
+      end)
+
+      describe('and backend returns 5XX', function()
+        it('does not invalidate the cache entry if there was a 4XX', function()
+          cache:set('a_key', 403)
+          ctx.cache_handler(cache, 'a_key', { status = 500 }, nil)
+          assert.equals(403, cache:get('a_key'))
+        end)
+
+        it('caches a 200 if there was nothing in the cache entry', function()
+          ctx.cache_handler(cache, 'a_key', { status = 500 }, nil)
+          assert.equals(200, cache:get('a_key'))
+        end)
+
+        it('caches a 200 if there was something != 4XX in the cache entry', function()
+          cache:set('a_key', 200)
+          ctx.cache_handler(cache, 'a_key', { status = 500 }, nil)
+          assert.equals(200, cache:get('a_key'))
+        end)
+      end)
+    end)
+
     describe('when disabled', function()
       local caching_policy
-      local cache
       local ctx
 
       setup(function()
@@ -99,7 +147,6 @@ describe('policy', function()
         caching_policy = require('apicast.policy.caching').new(config)
         ctx = {}
         caching_policy:rewrite(ctx)
-        cache = resty_lrucache.new(1)
       end)
 
       it('does not cache anything', function()